extracts metrics client
